src/requirements.txt
Flask
picamera
face_recognition
cv2
numpy
RPi.GPIO